frontend and backend

+++ backend/backend/Models/Rating.cs
@@ -28,8 +28,8 @@ namespace backend.Models
         public User User { get; set; }
 
         [Required]
-        [Range(1, 5, ErrorMessage = "Rating must be between 1 and 5.")]
-        public int Rate { get; set; }
+        [Range(1.0, 5.0, ErrorMessage = "Rating must be between 1 and 5.")]
+        public double Rate { get; set; }
 
         [MaxLength(1000)]
         public string Comment { get; set; }
